refactor(query-editor): split tab-query-editor into focused modules (#4)
First, low-risk pass at decomposing the 1,920-line tab-query-editor.tsx. Extracts
pure logic and self-contained UI/derivation into focused modules; no behavior change.
Container is now 1,568 lines.

- lib/table-query-builder.ts — pure generateWhereClause / generateOrderByClause /
  buildQueryWithFilters, now unit-tested (12 tests)
- query-editor/masked-export-dialog.tsx — the masked-export confirm AlertDialog
- query-editor/share-results-image.tsx — the share-results image render
- query-editor/use-editable-result.ts — getColumnsWithFKInfo / resolveEditSourceTable /
  getColumnsForEditing / getEditContext derivation hook

Removed the now-dead @data-peek/ui (AlertDialog family + cn), ShareImageDialog, and
several type imports.

Note: 3 pre-existing react-hooks/rules-of-hooks errors in this file (watch hooks after
an early return) are unchanged — left for a separate fix.

Verified: typecheck (node + web) clean, 862 tests pass, no new lint problems.
